// src/utils/enums/message-handler.ts
export enum MESSAGE_HANDLER {
  USER_NOT_FOUND = "User not found",
  EMAIL_ALREADY_TAKEN = "Email is already taken",
  USERNAME_ALREADY_TAKEN = "Username is already taken",
  INVALID_CREDENTIALS = "Email or password are incorrect",
  DOCUMENT_ALREADY_TAKEN = "The document number is already taken",
  CLIENT_NOT_FOUND = "The client was not found",
  CLIENT_CREATED = "The client was created successfully",
  CLIENT_UPDATED = "The client was updated successfully",
  CLIENT_DELETED = "The client was removed successfully",
  FORBIDDEN_RESOURCE = 'This action is not allowed this role.'
}